backend/scripts/requirements.txt
psycopg2==2.9.9      # PostgreSQL adapter for Python
bcrypt==4.1.2        # Password hashing
python-dotenv==1.0.1 # Environment variable loader
argparse==1.4.0      # Command-line argument parser
